refactor: add .repository to repository file names
