+++ jest.config.js
@@ -2,4 +2,10 @@
 module.exports = {
   preset: 'ts-jest',
   testEnvironment: 'node',
+  coveragePathIgnorePatterns: [
+    'node_modules',
+    'dist',
+    '.mock.ts',
+    '.repository.ts',
+  ],
 };
