+++ package.json
@@ -7,6 +7,7 @@
     "start": "./dist/index.js",
     "dev": "nodemon ./src/index.ts",
     "test": "jest",
+    "test:coverage": "jest --coverage",
     "lint": "eslint . --ext .ts --fix"
   },
   "keywords": [],
